requirements.txt
functions-framework
pandas
requests
pytz
google-cloud-bigquery
pyarrow
python-dotenv